dubai: Modify automatic brightness values
The values extracted from Stock ROM are quite dim on average and the minimum brightness is low,
So let's remove the three minimum light sensor values on config_autoBrightnessLevels and the
bottom four values and then add another value near max on config_autoBrightnessDisplayValuesNits
to help increase the minimum brightness and also bump up the values for the other automatic
brightness values.

While we're at it, we shall also raise the values for DarkeningLightDebounce and
BrighteningDebounce overlays so that it takes a lot longer before our device decides to actually
lower the brightness to prevent super jarring brightness changes.

Change-Id: Iacafa69d1e27b5201d98a832305c9afc5f228dd3